Save memory when resolving external objects.
This allows us to free the data for the original binary once we
have collected all of its external debug info reference information
and symbolicated non-external addresses.

In practice, this means that we will no longer have the data for
XUL (218.7MB) and libjs_static.a (710.5MB) in memory at the same time.