Use object instead of goblin for FatArch parsing.
I thought this would be the last dependency on goblin, but it turns out
there is another one: In pdb.rs, we use goblin to parse PE files and then
get their codeview_pdb70_debug_info, from which we need the PDB file path,
and the debug signature + age.
So we're still using goblin for that.