Allow partial section data reading during addr2line.
This works, but it doesn't seem to save much.
I've checked some of the benchmarks and we seem to be reading
parts of the data multiple times, so that the 'bytes read'
counter can go to over 100% for some files.
I haven't yet counted the percentage of 'unique bytes' accessed
(haven't found a RangeSet implementation that I like), so I can't
say whether we're just reading small parts over and over, or whether
we're actually reading most of the file.