Do a checked u64 -> usize conversion.
Wasm is 32 bit, so this conversion can truncate.
In some cases, this function was called with a huge size (due to bad input data in the file),
and then we'd truncate it to a reasonable size, and the read would unintentionally succeed.
But it wouldn't actually read the requested number of bytes.