Handle overflowing ranges better and avert panics early.
There was an integer overflow bug here which then resulted in a panic in range_map.
If the read range was completely outside the file size, we'd get an inverted
range with end < start because we'd only clamp the end of the range to the
file size, after rounding up to chunk boundaries. From this inverted range
we computed a read_len which would be huge due to wraparound during end-start.
Then we'd call insert_buffer_range with end < start and that panicked.