Add an experimental /sources/v1 entry point.
This allows obtaining source code for local files, but
only for files which are referenced by the debug symbol
files. The requester needs to bring proof in the form of
a library debugName+debugId and an address. Access is
only granted if the frames for that address refer to the
requested file.

This approach was suggested in
https://github.com/firefox-devtools/profiler/issues/3348#issuecomment-912991368