Use the new dyld split cache support from object.
This adds support for macOS 12 system library symbolication.
I had a copy of dyld_cache.rs in this repository which worked with
split caches on earlier macOS 12 Betas, but not on macOS 12 release.
In release, different sections of the same image can now be stored in
different SubCaches, so making this work required some deeper-reaching
changes to object.
This commit removes the non-working copy and uses the proper upstream
object support.