Hide symbolication API behind a randomly-generated path prefix.
The point of this is to make the URL harder to predict,
for the parts of the server that need to be accessible
with permissive CORS headers.

This is motivated by two things:
1. The /source/v1 route might be an attractive attack
vector to read arbitrary files. You cannot read arbitrary
files though, you can only read files whose paths are
referenced by (trusted) symbol files, and only if you know
the debugName + breakpadId for those symbol files, and have
an address which symbolicates to a frame that references the
file. Now, even if you know all those things, you additionally
need to know the randomly-generated path prefix.
2. In theory, different web pages could fetch http://127.0.0.1:3000/profile.json
and then track your identity across origins if they observe
the same profile. This is rather unlikely to begin with, but
not impossible. With this change, the profile.json is now
served at a different path every time profiler-symbol-server
is started, and the tracking vector is avoided.