Synthesize symbols from eh_frame info.
This works around the following problem:
Many libraries are distributed in a "stripped" form, i.e. with many of
their original symbols removed from the symbol table. For example, for
Firefox builds from the Release channel, libxul.so only has a few
symbols (mostly just exported "dynamic" symbols), so for most functions
we don't know the function names.
Now, if we want to look up the function name for an address, and we only
use the partial (stripped) symbol table, then it's likely that we
misattribute the address to the wrong function; we'd just blame the
closest function which still has a symbol.
In this case it would be much better to say "There is no symbol for this
function, even though there are some symbols for other functions in this
library". Or, alternatively, we could express "This address is inside
a function which starts at address abcd, but we don't know that function's
name."

This patch takes the latter approach: For functions without symbols, we
synthesize function names of the form fun_abcd at the function start
address. And we get those function start addresses (or at least an
approximation of them) from the unwinding information in the .eh_frame
section, which contains an FDE roughly for every function.
If we do have symbol names for the functions at these addresses, the
placeholder names are overwritten with the correct names.

This solves the ELF part of #23.