Use LC_FUNCTION_STARTS on macOS.
Now we use .eh_frame on Linux and LC_FUNCTION_STARTS on macOS to create
fallback symbols for stripped libraries.

The following command now creates 4376 fallback symbols on my machine:

cargo run --release -p dump-table -- AGXMetal13_3 /System/Library/Extensions/AGXMetal13_3.bundle/Contents/MacOS/

In the future I'd like to use these function start addresses in order to
compute a size for each function, so that we can return that size from
symbolication, so that the profiler can request the right range of bytes
when it wants to show the function's assembly code.