Also get function start addresses from __unwind_info.
In binaries which have code in both the __text section and in the
text_env section, it appears that LC_FUNCTION_STARTS only covers
functions in the __text section. But __unwind_info also has entries
for functions in the text_env section.
We don't want to use __unwind_info exclusively because __unwind_info
doesn't list all function start addresses: If there are two functions
with the same unwind opcode right next to each other, the two collapse
into one entry and the start address of the second function is lost.