Make startup faster.
To create the (debugName, debugId) -> path map, parse the JSON in a way
that lets serde discard any JSON subtrees we're not interested in. Before
this patch, we would heap allocate the entire object structure. Now we only
allocate the libs arrays (and their parent objects).

And we no longer keep the profile JSON bytes in memory; instead, they're
read again from disk when the /profile.json URL is requested.
If the file ends in .gz, then we forward it in gz form and set the
Content-Encoding: gzip header.

We also push the profile.json response out in 1MB chunks, with async
file reads.