Add a fxprof-processed-profile crate.
This crate is a sibling crate to the `gecko_profile` crate.
Using this crate can be more efficient because it allows the Firefox
Profiler to skip a processing step, and because it supports a "weight"
column in the sample table. The sample weight can be used to collapse
duplicate consecutive samples into one sample, which means that the
individual sample timestamps don't have to be serialized into the JSON.