Speed up serialization massively.
This commit switches perfrecord to use `fxprof-processed-profile` instead
of `gecko_profile` for generating the profile JSON.
This format supports sample weights, which allows us to save a ton of
space for zero-cpu-delta samples. In profiles with many threads, most of
the space in the JSON was taken up by the timestamps of samples from
idle threads. With the new format, we no longer emit individual
samples for consecutive zero-cpu-delta samples; instead, these idle
periods will be compressed down into one sample, and that sample will
have an appropriate "weight".
This makes the call tree, flame graph and stack chart look the same as
before (well, mostly [1]), but at a fraction of the space.

Here are some numbers. I took two recordings of around 42 seconds of
Firefox usage, before and after this change. Both profiles contained
around 1700 threads.

Serialization time went down from 2.5 seconds to 65ms. 40x improvement
The gz size of the profile went down from 160MB to 8MB. 20x improvement
Total opening time went down from 23 seconds to 3 seconds. 8x improvement

(Total opening time is the time from the "All tasks terminated" message
until the Firefox Profiler UI shows the threads and becomes interactive,
it does not include symbolication time.)

[1] The timelines look different now because we no longer have many
small squares underneath the graph for the individual sample times.
Clicking in those parts of the UI may not always select the targeted stack.
But the biggest difference is actually the behavior when selecting a
time range: The flame graph / call tree don't reflect partial selections
of idle periods. Instead, the weight of an idle period in the graph is
always "all or nothing", depending on whether the timestamp of the *end*
of that period is in the selected time range - the timestamp of the
weighted sample is set to the end of the idle period.
This behavior is quite unfortunate and should be fixed, but I'm not
really sure how. It is "correct behavior" in so far that it correctly
reflects the information that we write into the profile.
It would be more straightforward to fix if our weights were "durations"
rather than "sample weights", but using durations wouldn't be right
either, because we still want to display sample counts rather than times.
Maybe we can have two columns, i.e. both a weight and a duration column.