Emit event records sorted by time.
Improve how we deal with sample_id_all and add a CommonData type.
Add a way to get the timestamp of a record without parsing the entire record.
Rename a few things to make a distinction between "events" and "records".

In the perf.data files I've been testing with, only one "event" was
observed, which means that only one PerfEventAttr was in the attr list.
But many different types of records were emitted for that event.

I have not yet tested with a file which contains multiple events. I am
not sure how to match records to their event.