Use strings instead of paths for lib paths.
This crate is supposed to be usable on all platforms for creating
profiles recorded on any platform. For example, we want to be able
to generate profiles which were generated on Linux (with Linux paths)
while running the profile generator code on Windows. So we shouldn't
make the path syntax dependent on the current OS.