Only respect text sections when adding symbol end addresses.
This fixes the behavior on jitted-1234-56.so files which are produced by
; these .so files have bad section addresses which
cause all sections to overlap because they all start at address zero.

But we're only interested in sections with executable code anyway,
and these .so files only have one of those, so this seems like a
reasonable fix anyway.