Make samply-server build again, by ensuring the future is Send if FileContents is Sync.
The Uplooker is not Sync, only Send, because it contains an addr2line::Context which is only Send but not Sync.
This means that the future is allowed to contain an Uplooker but not a reference to an Uplooker.
Unfortunately, somehow the future was trying to contain a reference to an Uplooker.
I worked around it by making sure I don't need the Uplooker across an await point.

However, this is a bit concerning because it's easy to make a reference if you own an object.
Requiring the object to be Sync just because you have accidentally make a reference to it across an await point seems annoying.