Work around the Send problem.
We now force the FileContents trait to be Send + Sync if the `send_futures`
feature is enabled. Now the Send-ness of the future no longer depends on
the Send-ness of the concrete `FileContents` type.

This allows us to make the `ExternalFileWithUplooker` `Send`, with the
help of `Box<dyn ExternalFileUplookerTrait + 'a + Send + Sync>`. I had
to add a few more `#[cfg(feature = "send_futures")]` annotions for this.
Without the `Box<dyn>` we could have made the Send-ness dependent on the
Send-ness of the concrete `FileContents` type, but I needed the `Box<dyn>`
to make the Yoke "dependent" type covariant, because the concrete dependent
type uses interior mutability and thus isn't covariant.

Phrased differently: I have a self-referential struct, where the concrete
"dependent" type is invariant over the life time `'a`. With a `Box<dyn>`,
I can make a covariant wrapper around it. But this wrapper forces me to
sacrifice dynamic content-based `Send`-ness.
So I had to make the `Send`-ness fixed, based on a feature flag.

The feature is needed because I have both users which want `Send`, and
users which cannot provide `Send`:

 - samply-server uses a multithreaded tokio executor which requires
   `Send` futures
 - profiler-get-symbols-wasm implements `FileContents` and
   `FileAndPathHelper` with types which cannot be `Send` due to
   `wasm_bindgen`