Filter out symbols at address zero.
This was motivated by a test in fix-stacks which checks that
very low addresses don't symbolicate. We were picking up some
stray symbols with address zero for those.

It is known that .o files may legitimately contain functions at
address zero, but we don't create SymbolMaps for those, hopefully.
They should only be loaded as ExternalFiles.