Add proper GNU debug link support.
This is completely untested.

The old "fake" debug link support probably never worked, because
it would always try to validate the debug ID and fail; these binaries
probably don't have a GNU build ID, and the fallback computation would
find an empty .text section in the debug file.

Now we check the CRC instead, and we propagate the debug ID from the
original file to the SymbolMap of the debug link file.