Make /asm/v1 work for libcorecrypto.dylib on macOS 13.0.1.
I noticed a broken assembly view for functions from libcorecrypto.dylib in this
profile: https://share.firefox.dev/3Ctc2Lm

The reason turned out to be incorrect file offset information in the section
information of this image in the dyld shared cache. I filed FB11929250 about it.

We now get the machine code bytes from the segment, not from the section. I'm keeping
the section fallback so that this code still works on 'perf inject --jit' libraries
which only have sections and no segments.