Limit stack depth to 500 frames (macOS only).
If this is too short, please file an issue with an example profile.

This limit is intended to increase the number of profiles that can be
successfully analyzed, because they will no longer crash the UI.

This is macOS-only for now but we should probably do the same on Linux
eventually. It's just that, at the moment, stacks on Linux are already
quite limited due to the fact that we copy a fixed number of stack bytes
out from a kernel. But I haven't tested the graphviz example on Linux.

See https://github.com/mstange/samply/issues/13#issuecomment-1369401541
for how to generate a profile with deep stacks.