Turn this repo into a cargo workspace.
This makes it easier to run cargo check and cargo test on everything.

Furthermore, check Cargo.lock into the repo because samply is a binary.

I had to fix some code that was causing build failures now that the
features of a crate are shared between all workspace members: Because
wholesym turns on the `send_futures` feature of samply-api and
samply-symbols, any code that was conditional on that feature anywhere
is now compiled when it wasn't before.