Breaking change to /asm/v1: debugId -> breakpadId
Rename the field in the request JSON format from debugId to breakpadId.

This is consistent with the Library JSON object in the Gecko profile format
and the Firefox Profiler processed profile format.

It also makes it clear that the ID is given in the breakpad format, and
not in the lowercase-separated-by-dashes format.

I think it's fine to make this change now because the /asm/v1 endpoint
is not used by anything yet - the relevant Firefox Profiler UI pull
request hasn't shipped yet.
However, both samply 0.11 and Firefox 110 are shipping the previous
implementation which expected a debugId field, but those will just
return an error when the debugId field is missing, which can be treated
the same as if the endpoint was not supported at all.