Add an optional field to the /asm/v1 request JSON.
This adds a field called continueUntilFunctionEnd that is respected on
the request JSON for the /asm/v1 endpoint. If set to true, disassembly
will continue until the end of the function that startAddress is found in.
The specified size will act as a minimum length.

This is a workaround for the fact that the official Mozilla symbolication API
currently does not return function size information.