Implement a .symindex binary format.
This commit fixes all known memory usage and performance issues with
Breakpad sym files in wholesym. It improves the latency of common symbol
queries for large xul.sym files by 4x-6x in the cached case.

In one example with a 520MB xul.sym file, it improved the symbol query
time from 914ms to 202ms (4.5x, see below for profiles).

Issue I: Generating the index from the mmapped sym file was causing the
entire file to be resident in memory at the same time.
Issue II: The index was generated every time and not cached. This means
that the entire sym file had to be read on every call to load_symbol_map.
Issue III: Generating the index happened after download, taking an extra
400ms before symbols could be query.
Issue IV: Generating the index spent some time in hashmap code.

We now parse the file during sym file download, and store the index on
disk in a .symindex file, in a cache directory specified using
SymbolManagerConfig. This is done in a streaming fashion, i.e. we parse
the file as it comes from the net and don't load the entire file into
memory.

We also generate the symindex file for existing .sym files if it's
missing. This generation happens in wholesym, and it is also done in a
"streaming" fashion; we load individual chunks from the file and don't
use mmap when generating the index.

This fixes I, II and III:
 - Memory usage is much lower because index generation only ever looks
   at a small chunk of the file, and doesn't use mmap any more.
 - The index is only generated once. Subsequent loads of the same symbol
   map will find the existing symindex file, which means that only tiny
   amounts of the sym file even need to be loaded from disk: only the
   parts that are needed for the looked-up addresses.
 - Index generation is now performed during sym file download, saving
   those extra 400ms.

As a drive-by, I also eliminated the hashing that was showing up in
profiles. This was hashing the index for FILE and INLINE_ORIGIN entries.
But those indexes are ascending integers, so instead of a HashMap we now
use a Vec with binary search. I introduced a SortedVecBuilder which
skips sorting if it detects that the entries are already sorted in the
sym file (which they usually are).

Profile of querying a fresh sym file that needs to be downloaded:
before: https://share.firefox.dev/3H8CXOa , after: https://share.firefox.dev/3DhEMqW

Profile of querying a locally cached sym file without an index:
before: https://share.firefox.dev/3Y2jMMR , after: https://share.firefox.dev/4013565

Profile of querying a locally cached sym file with existing symindex:
https://share.firefox.dev/3j9jzbR

The fully cached case is improved 4.5x: 914ms -> 202ms.

The tests above were done by serving the xul.pdb symbol query for
re-symbolicating this profile: https://share.firefox.dev/3R3W5l3