Revert "Make compute_base_avma yet more lenient."
This reverts commit 9699e99dd32a9eeb7beb96f6521c1df2405606c6.

There are more bugs with the output of `perf inject --jit` than I knew
about at the time I wrote the commit message, and the fix doesn't work.
I'll reland it with a more accurate message once I've had a chance to
test it.

(The perf commit which added PT_LOAD caused another regression: It made
the synthesized MMAP records point to the wrong place, because the .text
section now moved by the phdr amount. A fix for this has landed but
hasn't made it into the current perf release on ubuntu, for example.)