Add PE mapping correlation hack for Linux.
It turns out that often it's not possible to determine a PE binary's on-disk
image directly, because their on disk representation is not page aligned and
needs to be copied to an anonymous page instead.

This commit implements a heuristic relying on the PE header always being at
the beginning of the file and mapped by Wine, to enable correlation of such
binaries.