Set special categories on V8 and Spidermonkey JIT functions.
This is used with perf.data files which were created by first recording d8 (the V8 shell)
or js (the Spidermonkey shell) with perf record, and then processing the perf.data file
with `perf inject --jit`.
Both JS engines add special prefixes to the symbol name to indicate the JIT tier.
With different categories, each of which has a different color, the different tiers
are now visible in the activity graph.