Mark JS JIT functions as JS.
This makes the JavaScript view in the Firefox profiler work better
on profiles which were generated with samply load from perf.data files
with JIT information.

Concretely, this is intended to help us compare the time spent in JS
functions across different JS engines. We can create two or three
profiles of the same workload run in different JS engines, and then
switch to the JS-only view to compare between them.