De-duplicate JIT functions in --merge-threads mode.
We de-duplicate based on name and code size.
We don't check the generated machine code for equality. That's because
we often have built-in pointers in the generated code which differ across
merged runs, and we'd still like to merge in this case.

This approach probably has some unintended false positives.