Always use the monotonic clock.
If you know of a reason not to do this, please let me know.

Using the monotonic clock means that the timestamps will be
comparable to timestamps taken in user space and written to
jitdump files.
With perf, you always have to add an extra flag if you want
to profile a program which makes use of jitdump. I would
prefer samply record to work with jitdump out-of-the-box.