Make the default category yellow on macOS.
This is consistent with what we were doing on Linux.

Yellow has better color contrast in the flamegraph against black text.
Unfortunately it has worse color contrast in the timeline activity graph.

Eventually I'd like the colors to be more dynamic, so that code from
different libraries or crates has different colors.