Linux: Use callback instead of iterators for perf event processing.
This transposes the perf event loop to an internal iteration-like
manner. The `consume_events` function calls the callback as long as
there are events to consume, then returns, at which the caller should
call into epoll to wait for new events.

This simplifies logic especially with the forthcoming live event
reordering.