Switch to SwCpuClock on ENOENT error
The `perf_event_open` manpage says that ENOENT is "Returned if the type
setting is not valid.". Prior to this fix, samply only switched to
SwCpuClick on an EPERM error. This fix means that we try it
speculatively whatever the error, unless it's a EPERM that can be fixed
by changing the paranoia level.