fxprof-processed-profile: add Profile::symbolicated
Previously it was emitted hardcoded to false. We now make it
settable so custom profile generators can emit symbolicated samples
without the Firefox Profiler attempting to incur extra work.

Note: there is also a `symbolicationNotSupported` ProfileMetadata
key. Perhaps it would be more appropriate to set instead or in addition
to? The Firefox Profiler docs are a bit unclear and I didn't look at
the source code to see how these interact.