fxprof-processed-profile: add initial selected/visible thread control
This gives custom profile producers more control over how threads are
displayed by default.

Because we sort the threads list on serialization, we have to supplement
the serialization code with a mechanism to translate the old thread
indices to the new ones. This extra complexity is somewhat annoying.
But I'm not sure what else to do.

This sorting code will likely get refactored eventually in order to
support the `keepProfileThreadOrder` ProfileMeta setting. I suppose
we'll deal with that later.