Add support for marker files on macOS.
If you open a file which starts with "marker-" and ends with ".txt", this
file will now be parsed and translated into markers on the main thread in the
profiler.

This is pretty much a hack and the format is unstable. For now the format is
as follows:
<start_timestamp> <end_timestamp> <marker_name>

It's a line-based text format. The timestamps are in nanoseconds relative to the
mach_absolute_time origin.