Look for the pdb right next to the binary file.
This helps, for example, when moving the build directory of a
Firefox Windows build to a different machine and profiling it there:
Inside xul.dll there is an absolute path to a pdb file, but this path is
only valid on the original machine - not on the machine where we're
symbolicating. So this path can't be used to locate the pdb immediately.
But the xul.pdb file is right next to the xul.dll file, so let's just
look there as well.