Add support for VDSO unwinding + symbolication.
The VDSO is a special ELF image that is mapped into
all processes on Linux. It is usually 4096 or 8192
bytes big. There's no file path for this mapping.
Instead, to get the bytes, we can look at the VDSO
mapping inside the samply process. The contents of
the VDSO are identical across all processes.

We need to handle the VDSO in three places:

 - For stack unwinding, when recording or when
   importing perf.data files with dwarf unwinding
 - For the LibraryInfo in the Profile, to store the
   right DebugId and CodeId in the profile
 - During symbolication, in case we don't find a
   debug symbols file for the VDSO, to fallback to
   getting symbols from the in-memory VDSO, as long
   as the DebugId matches