Provide a better error message when passing unknown arguments to samply record.
The first unrecognized argument will be treated as the command name.
If we can't launch an executable with that name, check whether this
argument might have been intended to be a flag.
Executables could legitimately start with a dash so we still launch
them if they exist, but we don't acknowledge this possibility in the
error message, to keep things simple in the common case.