Use "samply import perf.data" instead of "samply load perf.data".
Some command-line arguments, such as --profile-name and --reuse-threads,
only make sense when a new profile is being created. The "load" command
was trying to handle two different cases: displaying an existing JSON
profile in the profiler, and creating a new profile from a perf.data file.
The arguments noted above only made sense in the latter case.

So I've split the "load" command into two separate commands: "load" and
"import". The "load" command is now only for displaying an existing JSON
profile in the profiler. The "import" command is for creating a new profile
from a perf.data file, and you can now also specify the path of the
converted profile.