Use which.
Without this, runas didn't find xperf when I tried to run samply record
on my other Windows account. I'm not sure why. When I ran xperf directly
on the shell, it worked - the same shell that I used to launch samply.
I tried both on a regular cmd shell on the physical machine, and in
a VS Code ssh powershell (which runs as Administrator).