Improve handling of unmerged ETL files.
We still merge during recording. But this is one step on the way
to a world where we no longer need to do it. We could probably
stop merging right now for the case where we only have a kernel
session and no user session.

This commit tweaks how we buffer the ImageID information before we
process the MSNT_SystemTrace/Image information.
And it adds handling to get the CodeId + DebugId from the binary,
if we're importing an unmerged ETL file.

I've removed a fair amount of code that we had kept around for
adding kernel driver handling. It doesn't seem to be needed - I get
MSNT_SystemTrace/Image/DCStart events for all the same kernel modules
that I also see in the merged ETL.