Remove the concept of 'rounds' from the JitFunctionRecycler.
Instead, make functions available for recycling immediately.

Background:
A jitdump file may have multiple functions with the same name and the same code size.
For example, this can happen if the same jitted function gets recompiled with
different type information, or after it was GC'ed, or because the same script was
recompiled after a page navigation or page reload.
In the profiler assembly view, you may want to see the different assembly code for
the different recompilations.

When you want to collect more samples by running the workload multiple times,
and you relaunch the process, then each process creates its own Jitdump file.
We treat each Jitdump file as a 'library'. The Firefox Profiler creates separate
call nodes in the call tree for functions from separate libraries, even if those
functions have the same name. This means that, in order for the multi-iteration
multi-process --reuse-threads scenario to work properly, we somehow have to make
it so that function A from process X's jitdump file and function A from process
Y's jitdump file are treated as being from the same library. This is what the
JitFunctionRecycler is for: For function A from process Y, it will reuse the
information for function A from process X, making it look like both functions were
from the library we create for process X's jitdump file.

Now if you open the assembly view on such a JIT function, the sample count next
to each instruction will be an accumulated sample count over multiple rounds,
a combination of the sample counts from different processes and from different
compilations of the same function. The underlying gamble here is that each run
of the workload is similar enough that the generated assembly code is roughly the
same, modulo some built-in addresses and offsets.

So what happens if, within one process, you have multiple compilations of the
same JIT function with *different* generated assembly code, and they have the same
code size in bytes?
Before this commit, we would try to treat the separate compilations as separate
even in --reuse-threads mode, at least in the first process: The
JitFunctionRecycler would only try to match against functions from previous
processes, not from the current process.
After this commit, the JitFunctionRecycler will recycle functions even within the
same process. So you will no longer be able to see different assembly code for
recycled functions. But I think that's ok - the base assumption of --reuse-threads
is that the recompilations are similar enough that you don't need to see different
assembly code.